Alteração de arquitetura para página passar as funções e props ao componente
